App is finally working
Redid the UI so it's nicer, and fixed an issue with the BT where it would disconnect after a screen rotation due to a regular Activity calling onCreate() again and losing the reference to the previous BT connection.

This code also has numerous PreferenceActivity settings and a Wearable settings UI commented out because the settings are overly complicated and haven't been integrated with anything yet, and because the Wearable Settings UI is ugly, might not work, and that product isn't even out yet.